refactor(lite): 移除音频录制功能并优化扩展函数
- 删除了音频录制相关的 Fragment 和自定义视图组件
- 简化了 Bitmap 扩展函数，移除圆形图片生成功能
- 重构了广播管理器，改用单例模式并增强安全性
- 优化了 ByteArray 扩展函数命名和实现
- 移除了 Context 扩展函数中的 getSystemService 内联函数
- 优化了 Dialog 扩展函数中的属性访问方式
- 更新了文件操作相关扩展函数的实现方式
- 移除了界面布局中的音频录制相关组件
- 简化了上下文扩展函数中的系统服务获取方法

+++ app/src/main/java/com/pengxh/kt/lib/fragments/widget/DialogFragment.kt
@@ -173,7 +173,7 @@ class DialogFragment : KotlinBaseFragment<FragmentWidgetDialogBinding>() {
                         progressDialog.dismiss()
                     }
 
-                    override fun onProgressChanged(progress: Long) {
+                    override fun onProgressChanged(progress: Float) {
                         progressDialog.updateProgress(progress)
                     }
                 }).build().start()
